ai_gateway/requirements.txt
fastapi
uvicorn
openai
httpx
python-dotenv
pytest
pytest-asyncio
pytest-cov
structlog
opentelemetry-sdk
opentelemetry-exporter-otlp
